Restructuring files in preperation for reparenting building components.
