+++ Config/DefaultEngine.ini
@@ -82,6 +82,7 @@ bRemoteExecution=True
 +FunctionRedirects=(OldName="/Script/timbermvp.BuildSystemManagerComponent.VerticalSnapCondition",NewName="/Script/timbermvp.BuildSystemManagerComponent.SameOrientationSnapCondition")
 +FunctionRedirects=(OldName="/Script/timbermvp.TimberEnemyMeleeBase.CalculateDamage",NewName="/Script/timbermvp.TimberEnemyMeleeBase.CalculateOutputDamage")
 +PropertyRedirects=(OldName="/Script/timbermvp.TimberEnemyRangedBase.RangedWeapon",NewName="/Script/timbermvp.TimberEnemyRangedBase.EquippedWeapon")
++PropertyRedirects=(OldName="/Script/timbermvp.BuildSystemManagerComponent.ActiveBuildingComponent",NewName="/Script/timbermvp.BuildSystemManagerComponent.ActiveBuildingComponentProxy")
 
 [/Script/Engine.CollisionProfile]
 -Profiles=(Name="NoCollision",CollisionEnabled=NoCollision,ObjectTypeName="WorldStatic",CustomResponses=((Channel="Visibility",Response=ECR_Ignore),(Channel="Camera",Response=ECR_Ignore)),HelpMessage="No collision",bCanModify=False)
