+++ Content/StarterContent/Architecture/Floor_400x400.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:637d765368b81382e78608e2a1c4bebd44f808194316c71c0403aecffd3a295b
+oid sha256:d5b48e227dcb9baef7cb0f995f81a38e8e5a93d1053c1004ecd18d9f994769ad
 size 15406
