+++ Source/timbermvp/Private/BuildSystem/BuildingComponents/TimberBuildingComponentBase.cpp
@@ -1,7 +1,7 @@
 // Property of Paracosm Industries. Dont use my shit.
 
 
-#include "BuildSystem/TimberBuildingComponentBase.h"
+#include "BuildSystem/BuildingComponents/TimberBuildingComponentBase.h"
 
 #include "Character/Enemies/TimberEnemyCharacter.h"
 #include "Components/BoxComponent.h"
