+++ Source/timbermvp/Public/BuildSystem/BuildingComponents/TimberVerticalBuildingComponent.h
