+++ Source/timbermvp/Public/DataAssets/BuildComponentDataAsset.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "BuildSystem/TimberBuildingComponentBase.h"
+#include "BuildSystem/BuildingComponents/TimberBuildingComponentBase.h"
 #include "Engine/DataAsset.h"
 #include "BuildComponentDataAsset.generated.h"
